Add File, XMP finalizers
Note that this changes the old API in one significant way: now a
Fasttrack::XMP object created from an Exempi pointer is constructed
from a new copy of that pointer, not the identical pointer. The
Fasttrack version will be garbage collected, but the manual
Exempi pointer will not and you need to make sure you free it
yourself.